requirements.txt
streamlit
tensorflow
keras
numpy
h5py
pillow
opencv-python
matplotlib
seaborn

